refactor tests into suites.
This enables each implementation to subject itself to a common test
suite.
