+++ test/test_utils.go
@@ -1,4 +1,4 @@
-package testutil
+package test
 
 import (
 	"io"
